Refactoring

+++ src/test/java/es/upm/miw/apaw/theme/api/controllers/ThemeControllerIT.java
@@ -32,13 +32,13 @@ public class ThemeControllerIT {
         themeController.createTheme("tema1");
         new VoteController().createVote(1, 2);
         new VoteController().createVote(1, 3);
-        assertEquals(2.5, themeController.themeOverage(1).get(), 10e-2);
+        assertEquals(2.5, themeController.themeOverage(1).get().doubleValue(), 10e-2);
     }
 
     @Test
     public void testThemeOverageIfEmpty() {
         themeController.createTheme("tema1");
-        assertEquals(Double.NaN, themeController.themeOverage(1).get(), 10e-2);
+        assertEquals(Double.NaN, themeController.themeOverage(1).get().doubleValue(), 10e-2);
     }
 
     @Test
